add component ShowItem for show content. update component for ShowItem.
