+++ components/SignatureMetadataValidation.tsx
@@ -3,6 +3,7 @@ import { SignatureMetadata, AuthorSignatureMetadata } from '@metaio/meta-signatu
 import { verifyAuthorDigestMetadataSignature } from '@metaio/meta-signature-util';
 import { VerifyStatus } from '../utils/status';
 import VerifyResult from './VerifyResult';
+import ShowItem from './ShowItem';
 
 type SignatureMetadataValidationProps = {
   metadata: SignatureMetadata
@@ -25,22 +26,11 @@ const SignatureMetadataValidation = (props: SignatureMetadataValidationProps) =>
 
   return <div className="mt-8">
     <h2 className="font-thin text-2xl text-purple-700">Signature and validation</h2>
-    {
-      refer ? <div className="mt-2">
-        <p className="font-thin text-sm text-purple-700">Refer:</p>
-        <p className="p-1 my-1  font-thin text-xs bg-purple-100 rounded text-purple-700">{refer}</p>
-      </div> : <></>
-    }
-    {
-      sig ? <div className="mt-2">
-        <p className="font-thin text-sm text-purple-700">Signature:</p>
-        <p className="p-1 my-1 mt-2 font-thin text-xs bg-purple-100 rounded text-purple-700">{sig}</p> </div> : <></>
-    }
-    {
-      metadata.publicKey ? <div className="mt-2">
-        <p className="font-thin text-sm text-purple-700">Public key:</p>
-        <p className="p-1 my-1 mt-2 font-thin text-xs bg-purple-100 rounded text-purple-700">{metadata.publicKey}</p> </div> : <></>
-    }
+
+    {refer ? <ShowItem title="Refer:" content={refer} /> : <></>}
+    {sig ? <ShowItem title="Signature:" content={sig} /> : <></>}
+    {metadata.publicKey ? <ShowItem title="Public key:" content={metadata.publicKey} /> : <></>}
+
     <div className="flex flex-col mt-2">
       <p className="font-thin text-sm text-purple-700">Body:</p>
       <textarea value={customerMetaData} onChange={(e) => {
